+++ rtp_llm/cpp/normal_engine/NormalGenerateStream.cc
@@ -186,8 +186,8 @@ void NormalGenerateStream::updateOutput(const StreamUpdateInfo& update_info) {
     }
 
     // TODO: move it to better position
-    RTP_LLM_LOG_DEBUG("stream [%ld] finished: %d, pd_sep: %d, is_streaming: %d, need_remote_generate: %d",
-                      streamId(),
+    RTP_LLM_LOG_DEBUG("stream [%llu] finished: %d, pd_sep: %d, is_streaming: %d, need_remote_generate: %d",
+                      (unsigned long long)streamId(),
                       finished_,
                       queryPdSep(),
                       isStreaming(),
@@ -209,7 +209,7 @@ void NormalGenerateStream::updateOutput(const StreamUpdateInfo& update_info) {
         return;
     }
 
-    RTP_LLM_LOG_DEBUG("stream [%ld] enqueue generate output", streamId());
+    RTP_LLM_LOG_DEBUG("stream [%llu] enqueue generate output", (unsigned long long)streamId());
     enqueueGenerateOutput(prepareGenerateOutput(update_info));
 
     if (hasError()) {
